requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
Jinja2>=3.1.2
PyYAML>=6.0.0
pytest>=8.0.0
ruff>=0.8.0
